added long long int type, still doesn't resolve error though 	modified:   Elements/include/gambit/Elements/printable_types.hpp 	modified:   Printers/include/gambit/Printers/printers/hdf5printer.hpp

+++ Printers/include/gambit/Printers/printers/hdf5printer.hpp
@@ -310,8 +310,8 @@ namespace Gambit
 
         // Types compatible with the template print function
         #define TEMPLATE_TYPES      \
-         (int)(unsigned int)(long)(unsigned long)   \
-         (float)(double)
+         (int)(unsigned int)(long)(unsigned long)  \
+         (float)(double)(long long int) 
 
         // Types for which custom print functions are defined
         typedef std::map<std::string,double> map_str_dbl; // can't have commas in macro input
